upgrade the library to remove ng module
