+++ packages/ngx-clarity/src/index.ts
@@ -1,1 +1,1 @@
-export * from './lib/ngx-clarity.module';
+export * from './lib/ngx-clarity';
